+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Added
+- `extension-influxdb-metrics.rb` Include tag: metric and fields: issued, executed and received
 
 ## 0.1.1 - 2016-07-03
 ### Added